Move tests in a proper place
