--- api/controllers/base_controller_test.go
@@ -1,17 +1,0 @@
-package controllers
-
-import (
-	"reflect"
-	"testing"
-)
-
-func TestNewBaseController(t *testing.T) {
-	if reflect.TypeOf(NewBaseController()) != reflect.TypeOf(&BaseController{}) {
-		t.Errorf(
-			"%s return type is '%s' should be '%s'",
-			t.Name(),
-			reflect.TypeOf(NewBaseController()),
-			reflect.TypeOf(&BaseController{}),
-		)
-	}
-}
